version 13
Semantic versioning doesn't seem to make sense for binaries/apps. It's great for libraries though.

Past versions have not been consistent with semantic versioning at all and I can't seem to decide how to implement it for binaries/apps, so I've decided to drop it and just use the versionCode as the versionName.